+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 
+## [v0.1.2] - 2024-02-17
+### Added
+* Implement `delete{,_all}_{tasks,subroutines}`
+* Implement `begin_{subroutine,task}_chunk`
+
 ## [v0.1.1] - 2024-02-16
 ### Added
 * Increase send buffer size to accommodate all messages